GO-4969: Add UpdateSpaceRequest and UpdateSpaceHandler for modifying space details
