+++ core/api/model/space.go
@@ -9,6 +9,11 @@ type CreateSpaceRequest struct {
 	Description string `json:"description" example:"The local-first wiki"` // The description of the space
 }
 
+type UpdateSpaceRequest struct {
+	Name        string `json:"name" example:"New Space"`                   // The name of the space
+	Description string `json:"description" example:"The local-first wiki"` // The description of the space
+}
+
 type Space struct {
 	Object      string `json:"object" example:"space"`                                                                 // The data model of the object
 	Id          string `json:"id" example:"bafyreigyfkt6rbv24sbv5aq2hko3bhmv5xxlf22b4bypdu6j7hnphm3psq.23me69r569oi1"` // The id of the space
